# go-server-server/api/openapi.yaml
openapi: 3.0.3
info:
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger Petstore - OpenAPI 3.0
  version: 1.0.11
servers:
- url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /order:
    get:
      description: Returns list of orders
      operationId: getOrderList
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
      summary: Get list of all orders
      tags:
      - order
    post:
      description: Add a new order to the store
      operationId: addOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/xml:
            schema:
              $ref: '#/components/schemas/Order'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Order'
        description: Create a new order in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
          description: Successful operation
        "405":
          description: Invalid input
      summary: Add a new order to the store
      tags:
      - order
    put:
      description: Update an existing order by Id
      operationId: updateOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/xml:
            schema:
              $ref: '#/components/schemas/Order'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Order'
        description: Update an existent order in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        "405":
          description: Validation exception
      summary: Update an existing order
      tags:
      - order
  /order/{orderId}:
    delete:
      description: delete a order
      operationId: deleteOrder
      parameters:
      - description: Order id to delete
        explode: false
        in: path
        name: orderId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
          description: successful operation
        "400":
          description: Invalid order value
      summary: Deletes a order
      tags:
      - order
    get:
      description: Returns a single order
      operationId: getOrder
      parameters:
      - description: ID of order to return
        explode: false
        in: path
        name: orderId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
      summary: Find order by ID
      tags:
      - order
  /category:
    get:
      description: Returns list of categories
      operationId: getCategoryList
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Category'
                type: array
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Category not found
      summary: Get list of all categories
      tags:
      - category
    post:
      description: Add a new category to the store
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
          application/xml:
            schema:
              $ref: '#/components/schemas/Category'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Category'
        description: Create a new category in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: Successful operation
        "405":
          description: Invalid input
      summary: Add a new category to the store
      tags:
      - category
    put:
      description: Update an existing category by Id
      operationId: updateCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
          application/xml:
            schema:
              $ref: '#/components/schemas/Category'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Category'
        description: Update an existent category in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Category not found
        "405":
          description: Validation exception
      summary: Update an existing category
      tags:
      - category
  /category/{categoryId}:
    delete:
      description: delete a category
      operationId: deleteCategory
      parameters:
      - description: Category id to delete
        explode: false
        in: path
        name: categoryId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: successful operation
        "400":
          description: Invalid category value
      summary: Deletes a category
      tags:
      - category
    get:
      description: Returns a single category
      operationId: getCategory
      parameters:
      - description: ID of category to return
        explode: false
        in: path
        name: categoryId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Category not found
      summary: Find category by ID
      tags:
      - category
...
components:
...
  schemas:
    Order:
      example:
        quantity: 7
        id: 1
        shipDate: 2000-01-23T04:56:07.000+00:00
        complete: true
      properties:
        id:
          example: 1
          format: int64
          type: integer
        quantity:
          example: 7
          format: int32
          type: integer
        shipDate:
          format: date-time
          type: string
        complete:
          type: boolean
      type: object
      xml:
        name: order
    Category:
      example:
        name: Dog
        id: 1
      properties:
        id:
          example: 1
          format: int64
          type: integer
        name:
          example: Dog
          type: string
      type: object
      xml:
        name: category
...
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey