# go-server-server/go/api_order.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
 *
 * API version: 1.0.11
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// OrderApiController binds http requests to an api service and writes the service results to the http response
type OrderApiController struct {
	service OrderApiServicer
	errorHandler ErrorHandler
}

// OrderApiOption for how the controller is set up.
type OrderApiOption func(*OrderApiController)

// WithOrderApiErrorHandler inject ErrorHandler into controller
func WithOrderApiErrorHandler(h ErrorHandler) OrderApiOption {
	return func(c *OrderApiController) {
		c.errorHandler = h
	}
}

// NewOrderApiController creates a default api controller
func NewOrderApiController(s OrderApiServicer, opts ...OrderApiOption) Router {
	controller := &OrderApiController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the OrderApiController
func (c *OrderApiController) Routes() Routes {
	return Routes{ 
		{
			"AddOrder",
			strings.ToUpper("Post"),
			"/api/v3/order",
			c.AddOrder,
		},
		{
			"DeleteOrder",
			strings.ToUpper("Delete"),
			"/api/v3/order/{orderId}",
			c.DeleteOrder,
		},
		{
			"GetOrder",
			strings.ToUpper("Get"),
			"/api/v3/order/{orderId}",
			c.GetOrder,
		},
		{
			"GetOrderList",
			strings.ToUpper("Get"),
			"/api/v3/order",
			c.GetOrderList,
		},
		{
			"UpdateOrder",
			strings.ToUpper("Put"),
			"/api/v3/order",
			c.UpdateOrder,
		},
	}
}

// AddOrder - Add a new order to the store
func (c *OrderApiController) AddOrder(w http.ResponseWriter, r *http.Request) {
	orderParam := Order{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&orderParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertOrderRequired(orderParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.AddOrder(r.Context(), orderParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// DeleteOrder - Deletes a order
func (c *OrderApiController) DeleteOrder(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	orderIdParam, err := parseInt64Parameter(params["orderId"], true)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}

	result, err := c.service.DeleteOrder(r.Context(), orderIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetOrder - Find order by ID
func (c *OrderApiController) GetOrder(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	orderIdParam, err := parseInt64Parameter(params["orderId"], true)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}

	result, err := c.service.GetOrder(r.Context(), orderIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetOrderList - Get list of all orders
func (c *OrderApiController) GetOrderList(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetOrderList(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// UpdateOrder - Update an existing order
func (c *OrderApiController) UpdateOrder(w http.ResponseWriter, r *http.Request) {
	orderParam := Order{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&orderParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertOrderRequired(orderParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.UpdateOrder(r.Context(), orderParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}